moved utils fns
